+++ airbyte-webapp/src/packages/cloud/components/experiments/FreeConnectorProgram/EnrollmentButton/EnrollmentModal.tsx
@@ -1,16 +1,13 @@
 import React, { useState } from "react";
 import { FormattedMessage } from "react-intl";
 
-import { Button, ButtonVariant } from "components/ui/Button";
+import { Button } from "components/ui/Button";
 import { FlexContainer, FlexItem } from "components/ui/Flex";
 import { Heading } from "components/ui/Heading";
 import { ModalFooter } from "components/ui/Modal/ModalFooter";
 import { Text } from "components/ui/Text";
 
-import { useModalService } from "hooks/services/Modal";
 import { StripeCheckoutSessionCreate, StripeCheckoutSessionRead } from "packages/cloud/lib/domain/stripe";
-import { useStripeCheckout } from "packages/cloud/services/stripe/StripeService";
-import { useCurrentWorkspaceId } from "services/workspaces/WorkspacesService";
 
 import { CardSVG } from "./CardsSvg";
 import { ConnectorGridSvg } from "./ConnectorGridSvg";
@@ -26,7 +23,11 @@ interface EnrollmentModalContentProps {
   workspaceId: string;
 }
 
-const EnrollmentModalContent: React.FC<EnrollmentModalContentProps> = ({ closeModal, createCheckout, workspaceId }) => {
+export const EnrollmentModalContent: React.FC<EnrollmentModalContentProps> = ({
+  closeModal,
+  createCheckout,
+  workspaceId,
+}) => {
   const [isLoading, setIsLoading] = useState(false);
 
   const startStripeCheckout = async () => {
@@ -107,38 +108,5 @@ const EnrollmentModalContent: React.FC<EnrollmentModalContentProps> = ({ closeMo
         </FlexContainer>
       </ModalFooter>
     </>
-  );
-};
-
-interface FreeConnectorProgramEnrollmentButtonProps {
-  buttonTextKey?: string;
-  variant?: ButtonVariant;
-}
-
-export const FreeConnectorProgramEnrollmentButton: React.FC<FreeConnectorProgramEnrollmentButtonProps> = ({
-  buttonTextKey = "freeConnectorProgram.enrollmentModal.enrollButtonText",
-  variant = "primary",
-}) => {
-  const { openModal, closeModal } = useModalService();
-  const { mutateAsync: createCheckout } = useStripeCheckout();
-  const workspaceId = useCurrentWorkspaceId();
-
-  const showEnrollmentModal = () => {
-    openModal({
-      title: null,
-      content: () => (
-        <EnrollmentModalContent workspaceId={workspaceId} createCheckout={createCheckout} closeModal={closeModal} />
-      ),
-    });
-  };
-
-  return (
-    <FlexContainer justifyContent="flex-end">
-      <FlexItem>
-        <Button onClick={() => showEnrollmentModal()} variant={variant}>
-          <FormattedMessage id={buttonTextKey} />
-        </Button>
-      </FlexItem>
-    </FlexContainer>
   );
 };
