+++ airbyte-webapp/src/packages/cloud/components/experiments/FreeConnectorProgram/EnrollmentButton/FreeSvg.tsx
